# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delivery API Documentation",
        "description": "API para gestionar una flota de entregas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar un nuevo usuario/conductor",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "description": "Datos para el registro",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "document",
                                    "birth_date",
                                    "license_number",
                                    "phone"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan.perez@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-15"
                                    },
                                    "license_number": {
                                        "type": "string",
                                        "example": "B1-12345"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+573001234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente."
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "description": "Credenciales de usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso, devuelve el token de acceso."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada exitosamente."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/packages": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Listar los paquetes del conductor autenticado",
                "operationId": "2d2d4df67a3ce58700702ed1c7827acd",
                "responses": {
                    "200": {
                        "description": "Lista de paquetes paginada."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Packages"
                ],
                "summary": "Crear un nuevo paquete",
                "operationId": "f155b764ff7a75ce74b099a445398baa",
                "requestBody": {
                    "description": "Datos del nuevo paquete",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address",
                                    "package_status_id",
                                    "dimensions",
                                    "weight",
                                    "merchandise_type_id"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "Carrera 10 #20-30"
                                    },
                                    "package_status_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dimensions": {
                                        "type": "string",
                                        "example": "Medium"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "10 kg"
                                    },
                                    "merchandise_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paquete creado exitosamente."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/packages/{id}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Mostrar un paquete específico",
                "operationId": "3f922d3f0bd662184f000b78cb6ab115",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paquete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del paquete."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Actualizar un paquete",
                "operationId": "a33a5f0be34317d17d07c131ba5daacf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos a actualizar (pueden ser parciales)",
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paquete actualizado exitosamente."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "Eliminar un paquete",
                "operationId": "5e5d19ca2e0a8326156a56fd3d62b00b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paquete eliminado exitosamente."
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
            "Unauthorized": {
                "description": "No autenticado. Falta el token o es inválido.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Acceso denegado. No tiene permisos para acceder a este recurso."
            },
            "ValidationError": {
                "description": "Error de validación. Los datos enviados son incorrectos."
            },
            "NotFound": {
                "description": "El recurso solicitado no fue encontrado."
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por Token (Bearer)",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Packages",
            "description": "Packages"
        }
    ]
}